[ticket/13782] Rename null log to dummy log
A backwards compatible null class has been added that extends the dummy
log class.

PHPBB3-13782
