+++ tests/notification/group_request_test.php
@@ -52,7 +52,7 @@ class phpbb_notification_group_request_test extends phpbb_tests_notification_bas
 			$this->cache->get_driver()
 		));
 		$phpbb_dispatcher = new phpbb_mock_event_dispatcher;
-		$phpbb_log = new \phpbb\log\null();
+		$phpbb_log = new \phpbb\log\dummy();
 
 		// Now on to the actual test
